+++ lib/typecheck/typecheck.mli
@@ -147,6 +147,7 @@ type lin_entry = {
   le_first_use : Ast.span option ref;
   le_pending : ty option;
   le_dup : Ast.span option ref;
+  le_mixed : Ast.span option ref;
 }
 type ctor_info = {
   ci_type : string;